Still more work on skeletal mesh anims

+++ TinyFFR.Native/tffr/headers/scene/objects/native_impl_objects.h
@@ -6,7 +6,7 @@
 using namespace filament;
 using namespace filament::math;
 
-typedef int32_t ModelInstanceHandle;
+typedef uintptr_t ModelInstanceHandle;
 
 class native_impl_objects {
 public:
